move scp things into their own module
